requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
python-dotenv==1.0.1
aiofiles==23.2.1
Pillow==10.2.0
firebase-admin==6.4.0
pydantic==2.5.3
